+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
+- Fix crash caused by missing `SetForeground` api function
+
 ## [0.2.0] - 2025-10-28
 
 ### Added
@@ -46,8 +50,8 @@
 - First release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.0...HEAD
 
+[Unreleased]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/meehl/rusty-path-of-building/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/meehl/rusty-path-of-building/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/meehl/rusty-path-of-building/compare/v0.1.0...v0.1.1