+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
+- Fix version file
+
 ## [0.2.1] - 2025-10-28
 
 ### Added
@@ -56,8 +60,8 @@
 - First release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.1...HEAD
 
+[Unreleased]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/meehl/rusty-path-of-building/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/meehl/rusty-path-of-building/compare/v0.1.1...v0.1.2